Added continuity tests for rebasing. Also added some unit tests for various checks that are now modified to take account of baselines.

+++ Framework/Source/Baselines/CDERebaser.h
@@ -1,5 +1,5 @@
 //
-//  CDEBaselinePropagator.h
+//  CDERebaser.h
 //  Ensembles
 //
 //  Created by Drew McCormack on 05/01/14.
@@ -19,7 +19,7 @@
 
 - (instancetype)initWithEventStore:(CDEEventStore *)eventStore;
 
-- (CGFloat)estimatedEventStoreCompactionFollowingRebase;
+- (float)estimatedEventStoreCompactionFollowingRebase;
 - (BOOL)shouldRebase;
 
 - (void)rebaseWithCompletion:(CDECompletionBlock)completion;
